packages/hooke-px/pyproject.toml
[project]
name = "hooke-forge"
version = "0.1.0"
description = "The place where we forge all the hookes we need"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.11.0",
    "datasets>=4.4.1",
    "diffusers>=0.34.0",
    "imagecodecs>=2025.3.30",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "omegaconf>=2.3.0",
    "ornamentalist>=2.1.0",
    "polars>=1.31.0",
    "scanpy>=1.12",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.0",
    "submitit>=1.5.3",
    "timm>=1.0.17",
    "torch>=2.7.1",
    "torchdiffeq>=0.2.5",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "wandb>=0.21.0",
    "weave>=0.51.0",
    "hydra-core>=1.3.0",
    "hydra-zen>=0.13.0",
    "joblib>=1.4.0",
    "zarr>=3.1.5",
    # Note: txam package should be installed separately from local txam-inference repository
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hooke_forge"]

[tool.ruff]
exclude = ["*.ipynb"]
target-version = "py313"
line-length = 120
lint.select = ["E", "F", "W", "I", "UP"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.15.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[project.scripts]
hooke-train = "hooke_forge.training.train:cli"
hooke-train-tx-ae = "hooke_forge.training.train_tx_ae:cli"
hooke-eval = "hooke_forge.evaluation.eval:cli"
hooke-infer = "hooke_forge.inference.run:main"
hooke-gene-select = "hooke_forge.utils.tx_gene_selection:cli"
